requirements.txt
# Core Dependencies
yfinance==0.2.28
pandas
numpy<2.0

# Telegram Bot
python-telegram-bot==20.5

# Utilities
python-dateutil==2.8.2
pytz==2023.3
